Moving command code into their own folders
