+++ src/components/command-bar/commands/settings/dialog-body.tsx
@@ -1,0 +1,5 @@
+const SettingsCommandBarDialogBody = () => {
+	return <h1>Sutff for Settings command</h1>
+}
+
+export default SettingsCommandBarDialogBody
